requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
asgiref==3.5.0
    # via django
async-timeout==4.0.2
    # via redis
autopep8==1.6.0
    # via django-silk
billiard==3.6.4.0
    # via celery
cached-property==1.5.2
    # via kombu
celery==5.2.6
    # via
    #   -r requirements.in
    #   django-celery-results
cerberus==1.3.4
    # via -r requirements.in
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
coverage==6.3.2
    # via -r requirements.in
deprecated==1.2.13
    # via redis
django==3.2.13
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-redis
    #   django-silk
    #   djangorestframework
    #   drf-yasg
django-celery-results==2.3.1
    # via -r requirements.in
django-cors-headers==3.11.0
    # via -r requirements.in
django-grappelli==3.0.3
    # via -r requirements.in
django-redis==5.2.0
    # via -r requirements.in
django-silk==4.3.0
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   drf-yasg
drf-yasg==1.20.0
    # via -r requirements.in
gprof2dot==2021.2.21
    # via django-silk
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
importlib-metadata==4.11.3
    # via
    #   celery
    #   click
    #   kombu
    #   redis
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.1
    # via
    #   coreschema
    #   django-silk
kavenegar==1.1.2
    # via -r requirements.in
kombu==5.2.4
    # via celery
markupsafe==2.1.1
    # via jinja2
packaging==21.3
    # via
    #   drf-yasg
    #   redis
pillow==9.1.0
    # via -r requirements.in
prompt-toolkit==3.0.29
    # via click-repl
psycopg2-binary==2.9.3
    # via -r requirements.in
pycodestyle==2.8.0
    # via autopep8
pyparsing==3.0.8
    # via packaging
python-dateutil==2.8.2
    # via django-silk
python-decouple==3.6
    # via -r requirements.in
pytz==2022.1
    # via
    #   celery
    #   django
    #   django-silk
    #   djangorestframework
redis==4.2.2
    # via django-redis
requests==2.27.1
    # via
    #   coreapi
    #   django-silk
    #   kavenegar
ruamel.yaml==0.17.21
    # via drf-yasg
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sqlparse==0.4.2
    # via
    #   django
    #   django-silk
toml==0.10.2
    # via autopep8
typing-extensions==4.1.1
    # via
    #   asgiref
    #   async-timeout
    #   importlib-metadata
    #   redis
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.9
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
wrapt==1.14.0
    # via deprecated
zipp==3.8.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
